Move render-backend files into source/backends/
OpenGL backend implementations and the RenderBackend concept/selector
headers are relocated from source/ to source/backends/ so the directory
is the single home for all backend code (both window and render backends).

Files moved:
  source/{opengl_backend,opengl_mesh,opengl_texture,opengl_render_target}.{h,cpp}
  source/{render_backend,render_backend_select}.h
  → source/backends/

Include updates:
  canvas_impl.h/.cpp: render_backend_select.h → backends/render_backend_select.h

CMakeLists.txt source paths updated accordingly.
