+++ source/canvas_impl.cpp
@@ -10,7 +10,7 @@
 #include "mouse.h"
 #include "controller.h"
 #include "roboto_font.h"
-#include "render_backend_select.h"
+#include "backends/render_backend_select.h"
 #define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/matrix_transform_2d.hpp>
 #include <glm/ext.hpp>
