+++ source/canvas_impl.h
@@ -6,7 +6,7 @@
 #include "render_target_container.h"
 #include "texture_usage_monitor.h"
 #include "aa_box.h"
-#include "render_backend_select.h"
+#include "backends/render_backend_select.h"
 #include <vector>
 #include <utility>
